refactor code to give job results own component
